octave/symbolic: change test that WILL fail to be expected failure
The last test in /inst/vpa.m WILL fail at this point in time. This is
because of an underlying problem in Symbolic and the way it imports
sympy, which changed in sympy 1.6.

So, instead of having it be a dangerous warning, like the previous
one, we instead make it a test that we expect to fail, that way we
still have a record of the test having run and errored out, but the
result is propagated to the fntests.log file.