octave-packages: switch to using fetchurl from fetchTarball
fetchTarball is an IFD (import-from-derivation) which leads to a
download of the package DURING the evaluation of the source Nix file.

pkgs.fetchurl is the preferred way of handling the fetching of
tarballs from the internet, because that means we can specify what we
are supposed to get, and the download doesn't happen until we perform
the actual build.

octave-packages: switch to fetchurl from fetchTarball