octave-wrapper: build environment's site-wide octaverc file
octave-wrapper: unlink and copy site-wide octaverc file for editing

By default, octave will search for the environment variable
`OCTAVE_SITE_INITFILE` in
`/nix/store/hash-octave-version/share/octave/m/startup/octaverc`.
On a regular GNU/Linux distribution, this would be
`/usr/share/octave/m/startup/octaverc`.

However, we wrapped the various octave executables to have the
environment variable OCTAVE_SITE_INITFILE in our newly built
environment. So, we copy the original file over the environment (as
there may be configuration there we need to keep).

octave-wrapper: modify write-permission on site-wide octaverc

Because we are copying from the nix store, the file has read-only
permissions. However, we _need_ to be able to write to it, and as this
is a copy local to the environment, we chmod it to be able to write to
the file correctly.

octave-wrapper: give octave path of list of installed octave packages

This file is generated through the `pkg install` command from earlier
in the script. This file is an octave-generated file that is formatted
_FOR_ octave that lists the packages installed and certain metadata
about the packages.