octave-hook: add hook to write requiredOctavePackages to a file
This is heavily copied from Python's variant of
this (requiredPythonModules).

octave-packages.nix calls the default.nix in /hooks, which then writes
the script as a runnable script to the nix-store, and then runs it. It
is then sourced later.

All that happens is that the the list of required octave
packages (`requiredOctavePackages`) is written out to
required-octave-packages file in $out/nix-support.

This can be useful later when building the environment to figure out
exactly which version of a package we were referring to when this
package was built.

octave-hook: correct echo output for octave package building