octave: buildOctavePackage uses requiredOctavePackages for building
The `requiredOctavePackages` parameter is the one that the end-user
will specify in the package's .nix file.

The `requiredOctavePackages'` locally-bound value is the
recursively-computed set of ALL required octave packages that are
required for this package to work.

The `requiredOctavePackages` inside the `mkDerivation` is what is
actually used to build the package (and helps the end-user find out
the set of all required octave packages required for a given package
to work.

Lastly, we add the recursively-computer set of all required octave
packages to `buildInputs`, to ensure that they are also compiled and
available for installation into the end environment.