octave-env: use packages to install octave packages and wrap octave
First, install all packages that were both manually specified and were
required for the ones that were specified. Because each of the
required packages was built with the same buildOctavePackage function,
they will all contain the tarball required for installation.

Next, we need to wrap the octave programs with any environments both
the manually-specified octave packages required and any environments
required for any of the octave packages that were required by these ones.