octave.pkgs.ltfat: remove broken label by using patch file
octave.pkgs.ltfat: remove override with new buildOctavePackage

I completely redefined the buildOctavePackage function so that
patchPhases are readily available (because the package is now
unpacked). This removes the need for an override in this general case.
For more specific use-cases the end-user will still have to override
the package.