octave.pkgs.audio: remove broken label
Needed to add the jack2 and alsaLib libraries for rtmidi to be found
during the build. Unfortunately, that means that (for now) the audio
package is exclusive to Linux.

However, http://www.music.mcgill.ca/~gary/rtmidi/index.html#compiling
contains instructions for Mac OSX.
The dependencies for Mac are: CoreMIDI, CoreAudio, CoreFoundation