octave.pkgs: octave packages can have patchPhase and patches
This is an extension of the previous 2 commits.
Because the source is now unpacked, we can modify the source files
in-place, before re-tar-ing for the `pkg build` command.

This larger change means we do not need to override the package in its
definition to allow for modifications to the source before it runs.

octave-packages: make dontPatch false by default

Running through an empty patch phase is next-to-nothing in terms of
execution time, and doesn't affect the final result.

It just makes it easier for people to provide patching changes too.

octave-packages: set-union of package's patchPhase

This way, the default patchPhase is used when the programmer doesn't
specify anything. However, when they do, their patchPhase takes over
and is used.

octave.pkgs: package-building mkDrv inherits patchPhase info